Split days 12 through 23 into lib/main
